Fixed Units and folder locations
I read through the code and see that we are taking mean rate for each
of the values and the histogram.  Mean voltage is the circles in the
second and third figures, which I thought was useful because this will
help us see where sub threshold adaptation will like to be at as
current moves up.  We are in pA, since C is in pF.